src/ai/pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = ""
authors = ["eslupmi101 <ajsen-andreev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
fastapi = "0.95.0"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
